requirements.txt
# Flask ve Web Framework
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.1.1

# OpenAI & LLM
openai==1.13.3
langchain==0.1.15
langchain-openai==0.0.8
sentence-transformers==2.2.2
transformers==4.38.2
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
accelerate==0.26.1

# Web Scraping ve API Çağrıları
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.20.0
webdriver-manager==4.0.1

# JSON ve Veri İşleme
pandas==2.2.2
numpy==1.26.4
jsonlines==4.0.0

# Elasticsearch & TR Dizin
elasticsearch==8.13.0

# Vektör Veritabanı ve RAG
chromadb==0.4.24
faiss-cpu==1.7.4
pypika==0.48.9

# PDF ve Doküman İşleme
pymupdf==1.24.2
PyPDF2==3.0.1
pdfminer.six==20221105
unstructured==0.11.4

# Model Eğitimi ve Hugging Face
datasets==2.18.0
scikit-learn==1.4.2
protobuf==4.25.3

# Yardımcı Kütüphaneler
tqdm==4.66.2
python-dotenv==1.0.1
psutil==5.9.8
